comm: move BindingInspector to its consumer
